+++ Projects/2D_Roguelike/ProjectileManager.cpp
@@ -46,7 +46,7 @@ void ProjectileManager::Unload()
 
 void ProjectileManager::Update(float deltaTime)
 {
-	for (int i = 0; i < mProjectiles.size(); i++)
+	for (size_t i = 0; i < mProjectiles.size(); i++)
 	{
 		if (mProjectiles[i].IsActive())
 		{
@@ -57,7 +57,7 @@ void ProjectileManager::Update(float deltaTime)
 
 void ProjectileManager::Render()
 {
-	for (int i = 0; i < mProjectiles.size(); i++)
+	for (size_t i = 0; i < mProjectiles.size(); i++)
 	{
 		if (mProjectiles[i].IsActive())
 		{
@@ -68,7 +68,7 @@ void ProjectileManager::Render()
 
 void ProjectileManager::Fire(bool isPlayer,float speed,int damage, Math::Vector2 startingPos , Math::Vector2 targetPos)
 {
-	for (int i = 0; i < mProjectiles.size(); i++)
+	for (size_t i = 0; i < mProjectiles.size(); i++)
 	{
 		if (!mProjectiles[i].IsActive())
 		{
